requirement.txt
streamlit==1.35.0
opencv-python==4.10.0.82
Pillow==10.0.1
rembg==2.0.50
numpy==1.26.4
